Завершающая коррекция
