+++ index.html
@@ -31,9 +31,6 @@
 
       <div id="map" class="map__canvas"></div>
 
-      <!--<div  id="map-canvas"></div> Этот id почему-то неработал-->
-
-
       <!-- Фильтрация объявлений -->
       <div class="map__filters-container">
         <form action="#" class="map__filters" autocomplete="off">
@@ -57,7 +54,7 @@
             <option value="3">Три комнаты</option>
           </select>
           <select name="housing-guests" id="housing-guests" class="map__filter">
-            <option value="any" selected >Любое число гостей</option>
+            <option value="any" selected>Любое число гостей</option>
             <option value="2">Два гостя</option>
             <option value="1">Один гость</option>
             <option value="0">Не для гостей</option>
@@ -88,7 +85,8 @@
     <!-- Форма объявления -->
     <section class="notice">
       <h2 class="notice__title">Ваше объявление</h2>
-      <form class="ad-form" method="post" action = "https://22.javascript.pages.academy/keksobooking" enctype="multipart/form-data" autocomplete="off">
+      <form class="ad-form" method="post" action="https://22.javascript.pages.academy/keksobooking"
+        enctype="multipart/form-data" autocomplete="off">
         <fieldset class="ad-form-header">
           <legend class="ad-form-header__title">Ваша фотография (для карты)</legend>
           <div class="ad-form-header__upload">
@@ -96,7 +94,8 @@
               <img src="img/muffin-grey.svg" alt="Аватар пользователя" width="40" height="44">
             </div>
             <div class="ad-form__field">
-              <input type="file" id="avatar" name="avatar" class="ad-form-header__input visually-hidden" accept="image/png, image/jpeg">
+              <input type="file" id="avatar" name="avatar" class="ad-form-header__input visually-hidden"
+                accept="image/png, image/jpeg">
               <label class="ad-form-header__drop-zone" for="avatar">Загрузить<br>фото...</label>
             </div>
             <p class="ad-form-header__info">Заполните все обязательные поля, назначьте цену, загрузите аватар и фото
@@ -106,12 +105,13 @@
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="title">Заголовок объявления</label>
-          <input id="title" name="title" required type="text" placeholder="Милая, уютная квартирка в центре Токио" min="30" max="100">
+          <input id="title" name="title" required type="text" placeholder="Милая, уютная квартирка в центре Токио"
+            minlength="30" maxlength="100">
 
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="address">Адрес (координаты)</label>
-          <input id="address" name="address" type="text" readonly >
+          <input id="address" name="address" type="text" readonly>
         </fieldset>
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="type">Тип жилья</label>
@@ -151,9 +151,9 @@
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="capacity">Количество мест</label>
           <select id="capacity" name="capacity">
-            <option value="3" selected>для 3 гостей</option>
+            <option value="3">для 3 гостей</option>
             <option value="2">для 2 гостей</option>
-            <option value="1">для 1 гостя</option>
+            <option value="1" selected>для 1 гостя</option>
             <option value="0">не для гостей</option>
           </select>
         </fieldset>
@@ -187,7 +187,8 @@
           <label class="ad-form__label">Фотография жилья</label>
           <div class="ad-form__photo-container">
             <div class="ad-form__upload">
-              <input type="file" id="images" name="images" accept="image/png, image/jpeg" class="ad-form__input visually-hidden" >
+              <input type="file" id="images" name="images" accept="image/png, image/jpeg"
+                class="ad-form__input visually-hidden">
               <label for="images" class="ad-form__drop-zone">Загрузить<br>фото...</label>
             </div>
             <div class="ad-form__photo"></div>
