+++ js/main.js
@@ -6,4 +6,6 @@ import './map.js';
 import './form.js';
 import './avatar.js';
 import './filter.js';
+import './popups.js';
+
